added autorization

+++ micro/views.py
@@ -2,10 +2,20 @@ from flask import request, abort, make_response, jsonify
 
 from micro import app
 from micro.offers_api import get_offers
-from micro.api import add_product, update_product, delete_product, register_client
+from micro.api import add_product, update_product, delete_product, register_client,check_uuid
 
 ROUTE = '/micro/api/v1'
 """Routes for local API"""
+
+
+def authorized_access(response):
+    return check_uuid(response.headers.get('Bearer'))
+
+
+def process_response(res):
+    status_code = res['status_code']
+    del res['status_code']
+    return make_response(jsonify(res), status_code)
 
 
 @app.route(f'{ROUTE}/')
@@ -16,9 +26,7 @@ def home():
 @app.route(f'{ROUTE}/auth', methods=['POST'])
 def get_authorization():
     r = register_client()
-    status_code = r['status_code']
-    del r['status_code']
-    return make_response(jsonify(r), status_code)
+    return process_response(r)
 
 
 @app.route(f'{ROUTE}/<id>/offers')
@@ -28,25 +36,31 @@ def ll(id):
 
 @app.route(f'{ROUTE}/add', methods=['POST'])
 def add():
-    data = request.get_json()
-    r = add_product(data)
-    status_code = r['status_code']
-    del r['status_code']
-    return make_response(jsonify(r), status_code)
+    access, response = authorized_access(request)
+    if access:
+        data = request.get_json()
+        r = add_product(data)
+    else:
+        r = response
+    return process_response(r)
 
 
 @app.route(f'{ROUTE}/update', methods=['PUT'])
 def update():
-    data = request.get_json()
-    r = update_product(data)
-    status_code = r['status_code']
-    del r['status_code']
-    return make_response(jsonify(r), status_code)
+    access, response = authorized_access(request)
+    if access:
+        data = request.get_json()
+        r = update_product(data)
+    else:
+        r = response
+    return process_response(r)
 
 
 @app.route(f'{ROUTE}/<id>/delete', methods=['DELETE'])
 def delete(id):
-    r = delete_product(id)
-    status_code = r['status_code']
-    del r['status_code']
-    return make_response(jsonify(r), status_code)
+    access, response = authorized_access(request)
+    if access:
+        r = delete_product(id)
+    else:
+        r = response
+    return process_response(r)
